+++ chrisC.h
@@ -5,6 +5,6 @@
 #include <iostream>
 #include <cmath>
 //using namespace std;
-void calcChrisC(float r, float theta, float phi, float t, float chrisC1[4][4][4]);
+void calcChrisC(double r, double theta, double phi, double t, double chrisC1[4][4][4]);
 
 #endif
